add second test, flaky
